# Support SharedLister for NodeInfo