--- a/guest/api/framework.go
+++ b/guest/api/framework.go
@@ -1,0 +1,9 @@
+package api
+
+// ImageStateSummary provides summarized information about the state of an image.
+type ImageStateSummary struct {
+	// Size of the image
+	Size int64
+	// Used to track how many nodes have this image
+	NumNodes int
+}