--- a/guest/handle/sharedlister/internal/sharedlister.go
+++ b/guest/handle/sharedlister/internal/sharedlister.go
@@ -1,0 +1,11 @@
+package internal
+
+import guestapi "sigs.k8s.io/kube-scheduler-wasm-extension/guest/api"
+
+type SharedLister struct {
+	NodeInfoList guestapi.NodeInfoList
+}
+
+func (s SharedLister) NodeInfos() guestapi.NodeInfoList {
+	return s.NodeInfoList
+}