--- a/guest/handle/sharedlister/sharedlister_test.go
+++ b/guest/handle/sharedlister/sharedlister_test.go
@@ -1,0 +1,12 @@
+package sharedlister_test
+
+import (
+	"sigs.k8s.io/kube-scheduler-wasm-extension/guest/handle/sharedlister"
+)
+
+func ExampleNodeInfos() {
+	sharedlister.NodeInfos()
+
+	// Output:
+	//
+}