--- a/guest/postbind/postbind.go
+++ b/guest/postbind/postbind.go
@@ -65,7 +65,7 @@ func _postbind() { //nolint
 		return
 	}
 
-	nodeName := imports.NodeName()
+	nodeName := imports.CurrentNodeName()
 	// The parameters passed are lazy with regard to host functions. This means
 	// a no-op plugin should not have any unmarshal penalty.
 	postbind.PostBind(cyclestate.Values, cyclestate.Pod, nodeName)