--- a/guest/prebind/prebind.go
+++ b/guest/prebind/prebind.go
@@ -65,7 +65,7 @@ func _prebind() uint32 { //nolint
 		return 0
 	}
 
-	nodeName := imports.NodeName()
+	nodeName := imports.CurrentNodeName()
 	// The parameters passed are lazy with regard to host functions. This means
 	// a no-op plugin should not have any unmarshal penalty.
 	s := prebind.PreBind(cyclestate.Values, cyclestate.Pod, nodeName)