--- a/guest/prefilterextensions/prefilterextensions.go
+++ b/guest/prefilterextensions/prefilterextensions.go
@@ -21,10 +21,13 @@ package prefilterextensions
 
 import (
 	"sigs.k8s.io/kube-scheduler-wasm-extension/guest/api"
-	"sigs.k8s.io/kube-scheduler-wasm-extension/guest/filter"
+	"sigs.k8s.io/kube-scheduler-wasm-extension/guest/api/proto"
+	"sigs.k8s.io/kube-scheduler-wasm-extension/guest/handle/sharedlister"
 	"sigs.k8s.io/kube-scheduler-wasm-extension/guest/internal/cyclestate"
 	"sigs.k8s.io/kube-scheduler-wasm-extension/guest/internal/imports"
 	"sigs.k8s.io/kube-scheduler-wasm-extension/guest/internal/plugin"
+	internalproto "sigs.k8s.io/kube-scheduler-wasm-extension/guest/internal/proto"
+	protoapi "sigs.k8s.io/kube-scheduler-wasm-extension/kubernetes/proto/api"
 )
 
 // prefilterextensions is the current plugin assigned with SetPlugin.
@@ -55,7 +58,12 @@ func _addpod() uint32 { //nolint
 		return 0
 	}
 
-	status := prefilterextensions.AddPod(cyclestate.Values, cyclestate.Pod, &filter.PodInfo{}, &filter.NodeInfo{})
+	nodename := imports.CurrentNodeName()
+	if nodename == "" {
+		return imports.StatusToCode(&api.Status{Code: api.StatusCodeError, Reason: "could not get current node name"})
+	}
+
+	status := prefilterextensions.AddPod(cyclestate.Values, cyclestate.Pod, &podInfo{}, sharedlister.NodeInfos().Get(nodename))
 
 	return imports.StatusToCode(status)
 }
@@ -70,7 +78,66 @@ func _removepod() uint32 { //nolint
 		return 0
 	}
 
-	status := prefilterextensions.RemovePod(cyclestate.Values, cyclestate.Pod, &filter.PodInfo{}, &filter.NodeInfo{})
+	nodename := imports.CurrentNodeName()
+	if nodename == "" {
+		return imports.StatusToCode(&api.Status{Code: api.StatusCodeError, Reason: "could not get current node name"})
+	}
+
+	status := prefilterextensions.RemovePod(cyclestate.Values, cyclestate.Pod, &podInfo{}, sharedlister.NodeInfos().Get(nodename))
 
 	return imports.StatusToCode(status)
+}
+
+type podInfo struct {
+	pod proto.Pod
+}
+
+func (p *podInfo) GetApiVersion() string {
+	return p.lazyPod().GetApiVersion()
+}
+
+func (p *podInfo) GetKind() string {
+	return p.lazyPod().GetKind()
+}
+
+func (p *podInfo) GetName() string {
+	return p.lazyPod().GetName()
+}
+
+func (p *podInfo) GetNamespace() string {
+	return p.lazyPod().GetNamespace()
+}
+
+func (p *podInfo) GetResourceVersion() string {
+	return p.lazyPod().GetNamespace()
+}
+
+func (p *podInfo) GetUid() string {
+	return p.lazyPod().GetUid()
+}
+
+func (p *podInfo) Pod() proto.Pod {
+	return p.lazyPod()
+}
+
+func (p *podInfo) Spec() *protoapi.PodSpec {
+	return p.lazyPod().Spec()
+}
+
+func (p *podInfo) Status() *protoapi.PodStatus {
+	return p.lazyPod().Status()
+}
+
+// lazyPod lazy initializes pod from imports.Pod.
+func (p *podInfo) lazyPod() proto.Pod {
+	if pod := p.pod; pod != nil {
+		return pod
+	}
+
+	var msg protoapi.Pod
+	if err := imports.TargetPod(msg.UnmarshalVT); err != nil {
+		panic(err.Error())
+	}
+	p.pod = &internalproto.Pod{Msg: &msg}
+	return p.pod
 }