<!--  Thanks for sending a pull request!  Here are some tips for you:

1. If this is your first time, please read our contributor guidelines: https://git.k8s.io/community/contributors/guide/first-contribution.md#your-first-contribution and developer guide https://git.k8s.io/community/contributors/devel/development.md#development-guide
2. Please label this pull request according to what type of issue you are addressing, especially if this is a release targeted pull request. For reference on required PR/issue labels, read here:
https://git.k8s.io/community/contributors/devel/sig-release/release.md#issuepr-kind-label
3. Ensure you have added or ran the appropriate tests for your PR: https://git.k8s.io/community/contributors/devel/sig-testing/testing.md
4. If you want *faster* PR reviews, read how: https://git.k8s.io/community/contributors/guide/pull-requests.md#best-practices-for-faster-reviews
5. If the PR is unfinished, see how to mark it: https://git.k8s.io/community/contributors/guide/pull-requests.md#marking-unfinished-pull-requests
-->

#### What type of PR is this?

/kind feature
/kind cleanup

<!--
Add one of the following kinds:
/kind bug
/kind cleanup
/kind documentation
/kind feature

Optionally add one or more of the following kinds if applicable:
/kind api-change
/kind deprecation
/kind failing-test
/kind flake
/kind regression
-->

#### What this PR does / why we need it:

- Improve how the guest caches nodes for a better efficiency.
  - And, wire SharedLister().NodeInfos() and cached Nodes.
- Support SharedLister for NodeInfo.
  - Only the basic node's data and `ImageStates` is supported for now. We can support other fields later.
- Add [ImageLocality plugin](https://github.com/kubernetes/kubernetes/tree/master/pkg/scheduler/framework/plugins/imagelocality), ported from the in-tree plugin.

#### Which issue(s) this PR fixes:
<!--
*Automatically closes linked issue when PR is merged.
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
_If PR is about `failing-tests or flakes`, please post the related issues/tests in a comment and do not use `Fixes`_*
-->
Fixes #70 
Part of #73

#### Special notes for your reviewer:

#### Does this PR introduce a user-facing change?
<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required:
Enter your extended release note in the block below. If the PR requires additional action from users switching to the new release, include the string "action required".

For more information on release notes see: https://git.k8s.io/community/contributors/guide/release-notes.md
-->
```release-note
SharedLister for NodeInfo is supported.
```

#### What are the benchmark results of this change?
<!--
If documentation-only or trivial, just write "N/A". Otherwise, run the following:

1. `make bench-example > before.txt` on the commit prior to your changes
2. `make bench-example > after.txt` on the commit of your changes
3. Paste the output of `benchstat before.txt after.txt` as a code block.

If you haven't yet installed benchstat, it looks like this.
```
$ go install golang.org/x/perf/cmd/benchstat@latest
```

-->
```benchstat
                                       │  before.txt  │              after.txt               │
                                       │    sec/op    │    sec/op      vs base               │
Example_NodeNumber/Simple/New-12         563.2m ±  2%    584.6m ±  3%   +3.82% (p=0.015 n=6)
Example_NodeNumber/Simple/Run-12         93.41µ ± 26%   108.65µ ± 12%        ~ (p=0.093 n=6)
Example_NodeNumber/Simple_Log/New-12     552.2m ±  2%    614.9m ± 16%  +11.35% (p=0.002 n=6)
Example_NodeNumber/Simple_Log/Run-12     109.4µ ±  8%    108.9µ ± 28%        ~ (p=0.485 n=6)
Example_NodeNumber/Advanced/New-12       633.0m ±  3%    660.1m ±  3%   +4.28% (p=0.002 n=6)
Example_NodeNumber/Advanced/Run-12       37.80µ ± 25%    38.30µ ±  4%        ~ (p=0.589 n=6)
Example_NodeNumber/Advanced_Log/New-12   649.9m ± 12%    648.1m ±  5%        ~ (p=0.937 n=6)
Example_NodeNumber/Advanced_Log/Run-12   42.80µ ±  3%    44.93µ ± 10%   +4.98% (p=0.002 n=6)
geomean                                  6.175m          6.486m         +5.03%
```