@sanposhiho 2024-09-19
Merging the PR. 

@Gekko0114 (fyi @utam0k ) 
This repo is under development anyway and doesn't need to be conservative. So, let's not wait for other people's approval like k/k.
I mean not only PRs from me, but also ones from anyone. I know people besides me cannot `/approve` though, I'll basically just put the approval for PRs that either of you put the stamp.